Add dependencies Makefile generation to erlc(1) and compile(3)
This is useful when a project is built with Makefiles and erlc(1)
instead of EMakefiles. Tracking dependencies by hand is error-prone and
it becomes painful when using external application headers like EUnit's
one.

A dependencies Makefile will look like this:

    module.beam: module.erl \
      /usr/local/lib/erlang/lib/eunit-2.1.4/include/eunit.hrl \
      header.hrl

When included in the main Makefile, 'module' will be recompiled only
when needed.

GCC offers the same feature and new erlc(1) options are compatible with
it.

More informations at:
    http://wiki.github.com/dumbbell/otp/dependencies-makefile