Fix searching for source files in analyse_to_file/*
When looking for source files cover assumed they were in the same
directory as the .beam file or in ../src relative to that directory.
Now cover will first look at the location referenced in the .beam
before looking in the .beam directory then in relative ../src directory.

Updated tests for this new scenario.