beam_bool: Remove a clause in live_regs/1 that cannot match
live_regs/1 folds over a list of tuples, so a clause matching
an element being an empty list can never match.

If the list for some unfathomable reason would contain an empty
list, there will be an internal compiler error and no *.beam file
will be created. Thus this change is safe.